requirements.txt
arabic-reshaper==3.0.0
asgiref==3.6.0
asn1crypto==1.5.1
azure-core==1.30.1
azure-storage-blob==12.19.1
beautifulsoup4==4.12.2
bleach==6.0.0
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
crispy-bootstrap5==2023.10
cryptography==41.0.4
cssselect2==0.7.0
defusedxml==0.7.1
Django==4.2
django-allauth==0.57.0
django-autocomplete-light==3.9.7
django-autoslug==1.9.9
django-bootstrap-v5==1.0.11
django-cities-light==3.9.2
django-cors-headers==4.3.1
django-countries==7.5.1
django-crispy-forms==2.1
django-datatable-view==2.1.6
django-storages==1.14.2
django-summernote==0.8.20.0
html5lib==1.1
idna==3.4
isodate==0.6.1
lxml==4.9.3
oauthlib==3.2.2
oscrypto==1.3.0
Pillow==9.5.0
progressbar2==4.2.0
psycopg2==2.9.9
pycparser==2.21
pyHanko==0.20.1
pyhanko-certvalidator==0.24.1
PyJWT==2.8.0
pypdf==3.17.0
PyPDF2==3.0.1
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
python-decouple==3.8
python-utils==3.5.2
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
qrcode==7.4.2
reportlab==4.0.9
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
svglib==1.5.1
tinycss2==1.2.1
typing_extensions==4.10.0
tzdata==2023.3
tzlocal==5.2
Unidecode==1.3.6
uritools==4.0.2
urllib3==2.0.5
webencodings==0.5.1
xhtml2pdf==0.2.14
